\name{APERO-package}
\alias{APERO-package}
\alias{APERO}
\docType{package}
\title{
\packageTitle{APERO}
}
\description{
\packageDescription{APERO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{APERO}
\packageIndices{APERO}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{APERO}

Maintainer: \packageMaintainer{APERO}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
