% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apero.R
\name{apero}
\alias{apero}
\title{Creates Pandemic Apero}
\usage{
apero(room_length = 6, room_width = 4, min_distance = 1.5, margin = 0.3)
}
\arguments{
\item{room_length}{Length of room in meter.}

\item{room_width}{Width of room in meter.}

\item{min_distance}{Minimum distance in meter guests should have from each other.}

\item{margin}{How close to the walls do guests stand (in meter)?}
}
\value{
An object of class "apero" with the following elements:
\itemize{
\item \code{room_length}: User input.
\item \code{room_width}: User input.
\item \code{min_distance}: User input.
\item \code{margin}: User input.
\item \code{guests}: data.frame with x and y coordinates of guests (including yourself).
}
}
\description{
This function creates an apero in a room of given size.
Guests stick to a minimum distance of each other in order to stay safe.
}
\examples{
x <- apero()
x

x2 <- apero(1, 1)
plot(x2)
}
\seealso{
\code{\link[=plot.apero]{plot.apero()}}, \code{\link[=summary.apero]{summary.apero()}}.
}
