% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.apero}
\alias{print.apero}
\title{Prints an Apero}
\usage{
\method{print}{apero}(x, ...)
}
\arguments{
\item{x}{An object of class \code{apero}.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class \code{apero}.
}
\examples{
x <- apero()
x

x2 <- apero(1, 1)
x2
}
