% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.apero}
\alias{summary.apero}
\title{Summarizes an Apero}
\usage{
\method{summary}{apero}(object, ...)
}
\arguments{
\item{object}{An object of class \code{apero}.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the number of clinks is returned.
}
\description{
Summary method for an object of class \code{apero}.
}
\examples{
x <- apero()
summary(x)

x2 <- apero(2, 1)
summary(x2)
}
