% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.apero}
\alias{plot.apero}
\title{Plots an Apero}
\usage{
\method{plot}{apero}(x, emo = "beer", size = 8, color = "orange", ...)
}
\arguments{
\item{x}{An object of class \code{apero}.}

\item{emo}{An emoji, e.g., "beer", "drink", "syringe", "wine_glass".
Try emojifont::search_emoji("glass", approximate = TRUE) etc.}

\item{size}{Size of emoji.}

\item{color}{Color of emoji.}

\item{...}{Further arguments passed to \code{geom_text()}.}
}
\value{
A "ggplot" object.
}
\description{
Plot method for an object of class "apero".
}
\examples{
x <- apero()
x
plot(x)
plot(x, emo = "syringe", color = "red")

x2 <- apero(2, 1)
plot(x2)
}
